Resource files tweaked to support new UI designs and features
This change includes updates to the app icon, search icon, and information icon, with a new design to improve the consistency and visual appeal of the user interface.
In addition, the icons for message and contact actions have been updated to support dark themes and different color variables. Fixed color references in fab_new_message_bg.xml and ic_message.xml,
Make sure to be consistent with the latest UI specifications. The color values of the ic_checkmark_circle_blue icons have also been corrected to align with the colors used elsewhere.
```