''txt'' Permission changes for Android 13 (Tiramisu).
Introducing notification permission support for Android 13 (API level 33) and above. In newer versions of Android, apps must explicitly request permission to publish notifications. This update modifies the permission array in OsUtil and adds a new method to request the necessary notification permissions to ensure that the app is able to publish notifications on Android 13.

Changes include:
- Dynamically adjust permission arrays based on Android version.
- Added a new method 'notificationPermission() to return an array of appropriate notification permissions.
- Updated the 'hasNotificationPermission' method to reflect the use of the new permissions.
```

Signed-off-by: depth0 <5679381+depth0@user.noreply.gitee.com>